Improve CSS of Toolbar
