+++ olwebsite-app/src/OlWebsiteApp.js
@@ -4,11 +4,15 @@ import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 
 import "./style-sheets/main.scss";
 
-import Header from "./components/Header";
+import Header from "./components/Toolbar";
 import Footer from "./components/Footer";
 import LandingView from "./pages/LandingView";
 
 export default function OlWebsiteApp() {
+  /************************************
+   * Render
+   ************************************/
+
   return (
     <div className="app">
       <Router basename={process.env.PUBLIC_URL}>
