+++ olwebsite-app/src/components/Footer.js
@@ -1,5 +1,9 @@
 import React from "react";
 
 export default function Footer() {
+  /************************************
+   * Render
+   ************************************/
+
   return <footer></footer>;
 }
