+++ olwebsite-app/src/style-sheets/layout/_footer.scss
@@ -4,5 +4,4 @@ footer {
     flex-direction: row;
     align-items: center;
     min-height: 50px;
-    height: 250px;
 }
