--- olwebsite-app/src/style-sheets/layout/_header.scss
@@ -1,32 +1,0 @@
-header {
-    background-color: #ffffff;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-around;
-    align-items: center;
-    height: 60px;
-}
-
-.top-nav {
-    &-logo {
-    cursor: pointer;
-    height: calc(10px + 4vmin);
-    }
-
-    &-link {
-        border-bottom: 5px solid white;
-        cursor: pointer;
-        font-family: 'Open Sans', sans-serif;
-        height: 55px;
-        text-align: center;
-        vertical-align: middle;
-        line-height: 65px;
-        text-decoration: none;
-        color: black;
-    
-        &:hover, &:focus{
-            border-bottom: 5px solid #16341c;
-        }
-    }
-}
-
