+++ olwebsite-app/src/style-sheets/layout/_toolbar.scss
@@ -1,0 +1,68 @@
+.tool-bar {
+    background-color: #ffffff;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-around;
+    align-items: center;
+    height: 40px;
+}
+
+.top-nav {
+    &-logo {
+    cursor: pointer;
+    height: calc(8px + 4vmin);
+    }
+
+    &-link {
+        border-bottom: 5px solid white;
+        cursor: pointer;
+        font-family: 'Open Sans', sans-serif;
+        height: 35px;
+        text-align: center;
+        vertical-align: middle;
+        line-height: 45px;
+        text-decoration: none;
+        color: black;
+    
+        &-active {
+            border-bottom: 5px solid #16341c;
+        }
+
+        &:hover, &:focus{
+            border-bottom: 5px solid #16341c;
+        }
+    }
+
+    &-icon {
+        cursor: pointer;
+        font-family: 'Open Sans', sans-serif;
+        height: 35px;
+        text-align: center;
+        vertical-align: middle;
+        line-height: 43px;
+        text-decoration: none;
+        color: black;
+
+        &:hover{
+            opacity: .7;
+        }
+    }
+}
+
+.icon-container {
+    display: flex;
+    flex-direction: row;
+    min-width: 75px;
+    justify-content: space-around;
+}
+
+@media only screen and (max-width: 600px) {
+    /* For mobile: */
+    .icon-container {
+        min-width: 55px;
+    }
+
+    .top-nav-link {
+        display: none;
+    }
+  }
