+++ olwebsite-app/src/style-sheets/main.scss
@@ -1,6 +1,6 @@
 @import "./layout/app.scss";
 @import "./layout/footer";
-@import "./layout/header";
+@import "./layout/toolbar";
 
 @import "./components/carousel-wrapper";
